[FLINK-21998][hive] Move copied hive classes to a separate package
