+++ flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/planner/delegation/hive/DDLOperationConverter.java
@@ -80,6 +80,8 @@ import org.apache.flink.table.operations.ddl.DropPartitionsOperation;
 import org.apache.flink.table.operations.ddl.DropTableOperation;
 import org.apache.flink.table.operations.ddl.DropTempSystemFunctionOperation;
 import org.apache.flink.table.operations.ddl.DropViewOperation;
+import org.apache.flink.table.planner.delegation.hive.copy.HiveParserBaseSemanticAnalyzer;
+import org.apache.flink.table.planner.delegation.hive.copy.HiveParserStorageFormat;
 import org.apache.flink.table.planner.delegation.hive.desc.DropPartitionDesc;
 import org.apache.flink.table.planner.delegation.hive.desc.HiveParserAlterDatabaseDesc;
 import org.apache.flink.table.planner.delegation.hive.desc.HiveParserAlterTableDesc;
@@ -91,8 +93,6 @@ import org.apache.flink.table.planner.delegation.hive.desc.HiveParserDropDatabas
 import org.apache.flink.table.planner.delegation.hive.desc.HiveParserDropFunctionDesc;
 import org.apache.flink.table.planner.delegation.hive.desc.HiveParserDropTableDesc;
 import org.apache.flink.table.planner.delegation.hive.desc.HiveParserShowTablesDesc;
-import org.apache.flink.table.planner.delegation.hive.parse.HiveParserBaseSemanticAnalyzer;
-import org.apache.flink.table.planner.delegation.hive.parse.HiveParserStorageFormat;
 import org.apache.flink.table.planner.utils.OperationConverterUtils;
 
 import org.apache.hadoop.hive.metastore.api.FieldSchema;
