+++ flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/planner/delegation/hive/desc/HiveParserCreateViewDesc.java
@@ -18,7 +18,7 @@
 
 package org.apache.flink.table.planner.delegation.hive.desc;
 
-import org.apache.flink.table.planner.delegation.hive.parse.HiveParserASTNode;
+import org.apache.flink.table.planner.delegation.hive.copy.HiveParserASTNode;
 
 import org.apache.hadoop.hive.metastore.api.FieldSchema;
