+++ flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/planner/delegation/hive/parse/HiveASTHintParser.g
@@ -39,12 +39,13 @@ package org.apache.flink.table.planner.delegation.hive.parse;
 
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.hive.conf.HiveConf;
-import org.apache.flink.table.planner.delegation.hive.parse.HiveParserASTNode;
+import org.apache.flink.table.planner.delegation.hive.copy.HiveParserASTNode;
+import org.apache.flink.table.planner.delegation.hive.copy.HiveASTParseError;
 }
 
 
 @members {
-  ArrayList<HiveASTParseError> errors = new ArrayList<>();
+  public ArrayList<HiveASTParseError> errors = new ArrayList<>();
 
   @Override
   public void displayRecognitionError(String[] tokenNames,
