+++ flink-connectors/flink-connector-hive/src/main/java/org/apache/flink/table/planner/delegation/hive/parse/HiveParserDDLSemanticAnalyzer.java
@@ -24,10 +24,14 @@ import org.apache.flink.table.catalog.exceptions.CatalogException;
 import org.apache.flink.table.catalog.exceptions.PartitionNotExistException;
 import org.apache.flink.table.catalog.exceptions.TableNotExistException;
 import org.apache.flink.table.catalog.hive.HiveCatalog;
-import org.apache.flink.table.planner.delegation.hive.HiveParserAuthorizationParseUtils;
 import org.apache.flink.table.planner.delegation.hive.HiveParserConstants;
-import org.apache.flink.table.planner.delegation.hive.HiveParserContext;
-import org.apache.flink.table.planner.delegation.hive.HiveParserQueryState;
+import org.apache.flink.table.planner.delegation.hive.copy.HiveASTParseUtils;
+import org.apache.flink.table.planner.delegation.hive.copy.HiveParserASTNode;
+import org.apache.flink.table.planner.delegation.hive.copy.HiveParserAuthorizationParseUtils;
+import org.apache.flink.table.planner.delegation.hive.copy.HiveParserBaseSemanticAnalyzer;
+import org.apache.flink.table.planner.delegation.hive.copy.HiveParserContext;
+import org.apache.flink.table.planner.delegation.hive.copy.HiveParserQueryState;
+import org.apache.flink.table.planner.delegation.hive.copy.HiveParserStorageFormat;
 import org.apache.flink.table.planner.delegation.hive.desc.CreateTableASDesc;
 import org.apache.flink.table.planner.delegation.hive.desc.DropPartitionDesc;
 import org.apache.flink.table.planner.delegation.hive.desc.HiveParserAlterDatabaseDesc;
